pyproject.toml
[project]
name = "video-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.1.0",
    "daphne>=4.1.2",
    "django-channels>=0.7.0",
    "django>=5.1.1",
    "djangorestframework>=3.15.2",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
]
